Port methods from Calc to new Matrices class #558
